requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
protobuf==3.20.1
sqlparse==0.4.2
